# Implement fan control for Corsair HXi/RMi power supplies