--- a/docs/corsair-hxi-rmi.md
+++ b/docs/corsair-hxi-rmi.md
@@ -4,16 +4,16 @@
 
 ## Initialization
 
-It is necessary to initialize the device it has been powered on.
+It is necessary to initialize the device once it has been powered on.
 
 
 ```
 # liquidctl initialize
 ```
 
-The +12V rails normally function in multiple-rail mode, and `initialize` will set the PSU to that mode if necessary.  Single-rail mode can be selected by passing `--single-12v-ocp` to `initialize`.
+The +12V rails normally function in multiple-rail mode, and `initialize` will by default reset the PSU to that mode; single-rail mode can be selected by passing `--single-12v-ocp` to `initialize`.
 
-_Note: changing the +12V OCP mode is at the moment an experimental feature._
+_Note: changing the +12V OCP mode is currently an experimental feature._
 
 ## Monitoring
 
@@ -44,12 +44,10 @@ Total power                            110.00  W
 
 ## Fan speed
 
-The fan speed is normally controlled automatically by the PSU.
-
-It is possible to override this and set it to a fixed duty value using the `fan` channel.
+The fan speed is normally controlled automatically by the PSU.  It is possible to override this and set the fan to a fixed duty value using the `fan` channel.
 
 ```
 # liquidctl set fan speed 90
 ```
 
-This changes the fan mode to software control.  To revert back to hardware control, re-`initialize` the device.
+This changes the fan control mode to software control and sets the minimum allowed duty value to 30%.  To revert back to hardware control, re-`initialize` the device.