--- a/liquidctl.8
+++ b/liquidctl.8
@@ -254,7 +254,8 @@ back to hardware control, run \fBinitialize\fR again.
 .PP
 (Experimental feature) The +12V rails normally function in multiple-rail mode.
 Single-rail mode can be selected by passing \fB\-\-single\-12v\-ocp\fR to
-\fBinitialize\fR.
+\fBinitialize\fR.  To revert back to multitple-rail mode, run \fBinitialize\fR
+again without that flag.
 .
 .SS NZXT E500, E650, E850
 Fan channels: none (feature not supported yet).