Related: #45 ("Improve support for Corsair HXi/RMi PSUs")

Adds an implementation for `set_fixed_speed`, allowing fan control through the CLI:

```
# liquidctl set fan speed 35
```

The channel name is not currently enforced, but 'fan' seems like the obvious choice.  The duty cycle is so far only clamped to [0%, 100%].

---

_I still have a few questions..._

Does the code in this branch work?  
What is the minimum duty value for which the unit reacts?

```
# liquidctl initialize

# liquidctl status | grep 'Fan'

# liquidctl set fan speed 100
# liquidctl status | grep 'Fan'

# liquidctl set fan speed 0
# liquidctl status | grep 'Fan'

# liquidctl set fan speed 50
# liquidctl status | grep 'Fan'
...

[suggestion: do a binary search]
```

What happens to the fan speed when the unit doesn't receive constant updates?  Does it permanently stays the same (e.g. if left on 100% for a couple of minutes)?   Or does it revert back to some default behavior?

Presumably the unit must be able to not kill itself if iCue/CorsairLink has crashed.  One possible way to implement that would be for it to honor the specified duty as long as it is above some predefined minimum curve (based on temperature, load, or both).  Can we somehow observe this fallback behavior (e.g. by setting the fan speed to its minimum allowed value and ramping the power draw/temperature)?

Pinging @olielvewen for testing...

---

Missing:

 - [ ] clamping of duty to its minimum useful value
 - [ ] safety considerations