@olielvewen 2019-10-19
I just see it. For the moment, I m' looking for the man pages (in the git version I think) and after a feedback, I jump in this one. I can give you near 2 hours. Not sure for tomorrow under.......