@olielvewen 2019-10-19
> Does the code in this branch work?

No

> What is the minimum duty value for which the unit reacts?

If my memories is still good, I have said you the first time I have send you captures streams.
Normally 40% and not 35 %. Following the screenshots and the doc too. 

```
sudo liquidctl --version
liquidctl v1.2.0 (Manjaro Linux; python-liquidctl-git 1.2.0.r56.4e9de83-1)
]$ sudo liquidctl initialize
ERROR: Unexpected error
Traceback (most recent call last):
ValueError: Unknown format code 'x' for object of type 'str'
$ sudo liquidctl status --debug
[DEBUG] liquidctl.cli: liquidctl v1.2.0 (Manjaro Linux; python-liquidctl-git 1.2.0.r56.4e9de83-1)
Device 0, Corsair RM650i (experimental)
[DEBUG] liquidctl.driver.usb: replacing stock kernel driver with libusb
[DEBUG] liquidctl.driver.corsair_hid_psu: write 02 00 00 (and 61 padding bytes)
[DEBUG] liquidctl.driver.corsair_hid_psu: received 02 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
[DEBUG] liquidctl.driver.corsair_hid_psu: write 03 d2 (and 62 padding bytes)
[DEBUG] liquidctl.driver.corsair_hid_psu: received 03 d2 cf 58 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
[DEBUG] liquidctl.driver.corsair_hid_psu: write 03 d1 (and 62 padding bytes)
[DEBUG] liquidctl.driver.corsair_hid_psu: received 03 d1 9f c2 11 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
[DEBUG] liquidctl.driver.corsair_hid_psu: write 03 8d (and 62 padding bytes)
[DEBUG] liquidctl.driver.corsair_hid_psu: received 03 8d 96 f0 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
[DEBUG] liquidctl.driver.corsair_hid_psu: write 03 8e (and 62 padding bytes)
[DEBUG] liquidctl.driver.corsair_hid_psu: received 03 8e 7d f0 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
[DEBUG] liquidctl.driver.corsair_hid_psu: write 03 f0 (and 62 padding bytes)
[DEBUG] liquidctl.driver.corsair_hid_psu: received 03 f0 01 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
[DEBUG] liquidctl.driver.corsair_hid_psu: write 03 90 (and 62 padding bytes)
[DEBUG] liquidctl.driver.corsair_hid_psu: received 03 90 da 10 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
[DEBUG] liquidctl.driver.corsair_hid_psu: write 03 88 (and 62 padding bytes)
[DEBUG] liquidctl.driver.corsair_hid_psu: received 03 88 cc f9 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
[DEBUG] liquidctl.driver.corsair_hid_psu: write 03 ee (and 62 padding bytes)
[DEBUG] liquidctl.driver.corsair_hid_psu: received 03 ee 33 08 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
[DEBUG] liquidctl.driver.corsair_hid_psu: write 03 d8 (and 62 padding bytes)
[DEBUG] liquidctl.driver.corsair_hid_psu: received 03 d8 02 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
[DEBUG] liquidctl.driver.corsair_hid_psu: write 02 00 00 (and 61 padding bytes)
[DEBUG] liquidctl.driver.corsair_hid_psu: received 02 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
[DEBUG] liquidctl.driver.corsair_hid_psu: write 03 8b (and 62 padding bytes)
[DEBUG] liquidctl.driver.corsair_hid_psu: received 03 8b 08 d3 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
[DEBUG] liquidctl.driver.corsair_hid_psu: write 03 8c (and 62 padding bytes)
[DEBUG] liquidctl.driver.corsair_hid_psu: received 03 8c 1b f0 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
[DEBUG] liquidctl.driver.corsair_hid_psu: write 03 96 (and 62 padding bytes)
[DEBUG] liquidctl.driver.corsair_hid_psu: received 03 96 28 08 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 0
```