@olielvewen 2019-10-19
```
sudo liquidctl status | grep 'Fan' --debug
grep : option non reconnue '--debug'

Usage: grep [OPTION]... PATTERNS [FILE]...
Exécutez « grep --help » pour obtenir des renseignements complémentaires.
WARNING: reading the +12V OCP mode is an experimental feature
Exception ignored in: <_io.TextIOWrapper name='<stdout>' mode='w' encoding='UTF-8'>
BrokenPipeError: [Errno 32] Broken pipe
```