@olielvewen 2019-10-19
AH

```
sudo liquidctl status | grep 'Fan'
WARNING: reading the +12V OCP mode is an experimental feature
Fan control mode                 Software  
Fan speed                             876  rpm
```