@jonasmalacofilho 2019-10-19
> ```
> $ sudo liquidctl initialize
> ERROR: Unexpected error
> Traceback (most recent call last):
> ValueError: Unknown format code 'x' for object of type 'str'
> ```

Please send generate a stack trace for that error by running initialize again, this time with `--debug`.

> ```
> $ sudo liquidctl status --debug
> <only debug output>
> ```

Where's the rest/normal output?

> If my memories is still good, I have said you the first time I have send you captures streams.
Normally 40% and not 35 %. Following the screenshots and the doc too.

What iCue allows is not necessarily the same as what the PSU supports.  Because of this I prefer to figure out the actual minimum duty value empirically.

That said, if it ends up being a much lower than what iCue allows, we can add a safety note to the documentation.