@olielvewen 2019-10-19
>Please send generate a stack trace for that error by running initialize again, this time with --debug.

    $ sudo liquidctl status --debug
    <only debug output>

See 7 eme lines after liquidctl --version