@jonasmalacofilho 2019-10-19
> Please send generate a stack trace for that error by running initialize again, this time with --debug.

This is about the error during `initialize`, not `status`.