@olielvewen 2019-10-19
See in the folder rm650i screenshot opencorsairlinkusage-5.png and in the conversation...