@olielvewen 2019-10-20
```
sudo liquidctl initialize && sudo liquidctl set fan speed 30 --verbose && sleep 5 && sudo liquidctl status
INFO: ensuring fan control is in software mode
INFO: setting fan PWM duty to 30%
Device 0, Corsair RM650i (experimental)
WARNING: reading the +12V OCP mode is an experimental feature
Current uptime                    0:40:34  
Total uptime            13 days, 13:40:34  
Temperature 1                        37.8  °C
Temperature 2                        31.0  °C
Fan control mode                 Software  
Fan speed                             540  rpm
Input voltage                      230.00  V
Total power                        100.00  W
+12V OCP mode                  Multi rail  
+12V output voltage                 12.14  V
+12V output current                  6.75  A
+12V output power                   80.00  W
+5V output voltage                   4.97  V
+5V output current                   3.31  A
+5V output power                    16.00  W
+3.3V output voltage                 3.33  V
+3.3V output current                 1.62  A
+3.3V output power                   5.00  W
```
```mprime

Welcome to GIMPS, the hunt for huge prime numbers.  You will be asked a
few simple questions and then the program will contact the primenet server
to get some work for your computer.  Good luck!

Attention OVERCLOCKERS!!  Mprime has gained a reputation as a useful
stress testing tool for people that enjoy pushing their hardware to the
limit.  You are more than welcome to use this software for that purpose. 
Please select the stress testing choice below to avoid interfering with
the PrimeNet server.  Use the Options/Torture Test menu choice for your
stress tests.  Also, read the stress.txt file.

If you want to both join GIMPS and run stress tests, then Join GIMPS and
answer the questions.  After the server gets some work for you, stop
mprime, then run mprime -m and choose Options/Torture Test.

Join Gimps? (Y=Yes, N=Just stress testing) (Y): n
Number of torture test threads to run (16): 16
Choose a type of torture test to run.
  1 = Small FFTs (maximum heat and FPU stress, data fits in L2 cache, RAM
not tested much).
  2 = In-place large FFTs (maximum power consumption, some RAM tested).
  3 = Blend (tests some of everything, lots of RAM tested).
  11,12,13 = Allows you to fine tune the above three selections.
Blend is the default.  NOTE: if you fail the blend test, but can pass the
small FFT test then your problem is likely bad memory or a bad memory
controller.
Type of torture test to run (3): 1

Accept the answers above? (Y): y
[Main thread Oct 20 11:28] Starting workers.
[Worker #2 Oct 20 11:28] Worker starting
[Worker #3 Oct 20 11:28] Worker starting
[Worker #9 Oct 20 11:28] Worker starting
[Worker #5 Oct 20 11:28] Worker starting
[Worker #6 Oct 20 11:28] Worker starting
[Worker #7 Oct 20 11:28] Worker starting
[Worker #2 Oct 20 11:28] Beginning a continuous self-test on your computer.
[Worker #2 Oct 20 11:28] Please read stress.txt.  Hit ^C to end this test.
[Worker #8 Oct 20 11:28] Worker starting
[Worker #11 Oct 20 11:28] Worker starting
[Worker #10 Oct 20 11:28] Worker starting
[Worker #3 Oct 20 11:28] Beginning a continuous self-test on your computer.
[Worker #3 Oct 20 11:28] Please read stress.txt.  Hit ^C to end this test.
[Worker #10 Oct 20 11:28] Beginning a continuous self-test on your computer.
[Worker #10 Oct 20 11:28] Please read stress.txt.  Hit ^C to end this test.
[Worker #15 Oct 20 11:28] Worker starting
[Worker #1 Oct 20 11:28] Worker starting
[Worker #9 Oct 20 11:28] Beginning a continuous self-test on your computer.
[Worker #9 Oct 20 11:28] Please read stress.txt.  Hit ^C to end this test.
[Worker #5 Oct 20 11:28] Beginning a continuous self-test on your computer.
[Worker #5 Oct 20 11:28] Please read stress.txt.  Hit ^C to end this test.
[Worker #2 Oct 20 11:28] Test 1, 720000 Lucas-Lehmer iterations of M580673 using FMA3 FFT length 28K, Pass1=448, Pass2=64, clm=2.
[Worker #6 Oct 20 11:28] Beginning a continuous self-test on your computer.
[Worker #6 Oct 20 11:28] Please read stress.txt.  Hit ^C to end this test.
[Worker #7 Oct 20 11:28] Beginning a continuous self-test on your computer.
[Worker #7 Oct 20 11:28] Please read stress.txt.  Hit ^C to end this test.
[Worker #8 Oct 20 11:28] Beginning a continuous self-test on your computer.
[Worker #8 Oct 20 11:28] Please read stress.txt.  Hit ^C to end this test.
[Worker #13 Oct 20 11:28] Worker starting
[Worker #14 Oct 20 11:28] Worker starting
[Worker #12 Oct 20 11:28] Worker starting
[Worker #15 Oct 20 11:28] Beginning a continuous self-test on your computer.
[Worker #12 Oct 20 11:28] Beginning a continuous self-test on your computer.
[Worker #12 Oct 20 11:28] Please read stress.txt.  Hit ^C to end this test.
[Worker #4 Oct 20 11:28] Worker starting
[Worker #11 Oct 20 11:28] Beginning a continuous self-test on your computer.
[Worker #11 Oct 20 11:28] Please read stress.txt.  Hit ^C to end this test.
[Worker #14 Oct 20 11:28] Beginning a continuous self-test on your computer.
[Worker #14 Oct 20 11:28] Please read stress.txt.  Hit ^C to end this test.
[Worker #15 Oct 20 11:28] Please read stress.txt.  Hit ^C to end this test.
[Worker #16 Oct 20 11:28] Worker starting
[Worker #1 Oct 20 11:28] Beginning a continuous self-test on your computer.
[Worker #1 Oct 20 11:28] Please read stress.txt.  Hit ^C to end this test.
[Worker #13 Oct 20 11:28] Beginning a continuous self-test on your computer.
[Worker #13 Oct 20 11:28] Please read stress.txt.  Hit ^C to end this test.
[Worker #10 Oct 20 11:28] Test 1, 720000 Lucas-Lehmer iterations of M580673 using FMA3 FFT length 28K, Pass1=448, Pass2=64, clm=2.
[Worker #16 Oct 20 11:28] Beginning a continuous self-test on your computer.
[Worker #16 Oct 20 11:28] Please read stress.txt.  Hit ^C to end this test.
[Worker #9 Oct 20 11:28] Test 1, 720000 Lucas-Lehmer iterations of M580673 using FMA3 FFT length 28K, Pass1=448, Pass2=64, clm=2.
[Worker #4 Oct 20 11:28] Beginning a continuous self-test on your computer.
[Worker #4 Oct 20 11:28] Please read stress.txt.  Hit ^C to end this test.
[Worker #3 Oct 20 11:28] Test 1, 720000 Lucas-Lehmer iterations of M580673 using FMA3 FFT length 28K, Pass1=448, Pass2=64, clm=2.
[Worker #5 Oct 20 11:28] Test 1, 720000 Lucas-Lehmer iterations of M580673 using FMA3 FFT length 28K, Pass1=448, Pass2=64, clm=2.
[Worker #7 Oct 20 11:28] Test 1, 720000 Lucas-Lehmer iterations of M580673 using FMA3 FFT length 28K, Pass1=448, Pass2=64, clm=2.
[Worker #8 Oct 20 11:28] Test 1, 720000 Lucas-Lehmer iterations of M580673 using FMA3 FFT length 28K, Pass1=448, Pass2=64, clm=2.
[Worker #1 Oct 20 11:28] Test 1, 720000 Lucas-Lehmer iterations of M580673 using FMA3 FFT length 28K, Pass1=448, Pass2=64, clm=2.
[Worker #15 Oct 20 11:28] Test 1, 720000 Lucas-Lehmer iterations of M580673 using FMA3 FFT length 28K, Pass1=448, Pass2=64, clm=2.
[Worker #14 Oct 20 11:28] Test 1, 720000 Lucas-Lehmer iterations of M580673 using FMA3 FFT length 28K, Pass1=448, Pass2=64, clm=2.
[Worker #16 Oct 20 11:28] Test 1, 720000 Lucas-Lehmer iterations of M580673 using FMA3 FFT length 28K, Pass1=448, Pass2=64, clm=2.
[Worker #13 Oct 20 11:28] Test 1, 720000 Lucas-Lehmer iterations of M580673 using FMA3 FFT length 28K, Pass1=448, Pass2=64, clm=2.
[Worker #6 Oct 20 11:28] Test 1, 720000 Lucas-Lehmer iterations of M580673 using FMA3 FFT length 28K, Pass1=448, Pass2=64, clm=2.
[Worker #12 Oct 20 11:28] Test 1, 720000 Lucas-Lehmer iterations of M580673 using FMA3 FFT length 28K, Pass1=448, Pass2=64, clm=2.
[Worker #11 Oct 20 11:28] Test 1, 720000 Lucas-Lehmer iterations of M580673 using FMA3 FFT length 28K, Pass1=448, Pass2=64, clm=2.
[Worker #4 Oct 20 11:28] Test 1, 720000 Lucas-Lehmer iterations of M580673 using FMA3 FFT length 28K, Pass1=448, Pass2=64, clm=2.
[Worker #15 Oct 20 11:31] Self-test 28K passed!
[Worker #15 Oct 20 11:31] Test 1, 3200000 Lucas-Lehmer iterations of M172031 using FMA3 FFT length 8K, Pass1=128, Pass2=64, clm=2.
[Worker #12 Oct 20 11:31] Self-test 28K passed!
[Worker #12 Oct 20 11:31] Test 1, 3200000 Lucas-Lehmer iterations of M172031 using FMA3 FFT length 8K, Pass1=128, Pass2=64, clm=2.
[Worker #11 Oct 20 11:31] Self-test 28K passed!
[Worker #11 Oct 20 11:31] Test 1, 3200000 Lucas-Lehmer iterations of M172031 using FMA3 FFT length 8K, Pass1=128, Pass2=64, clm=2.
[Worker #2 Oct 20 11:31] Self-test 28K passed!
[Worker #2 Oct 20 11:31] Test 1, 3200000 Lucas-Lehmer iterations of M172031 using FMA3 FFT length 8K, Pass1=128, Pass2=64, clm=2.
[Worker #1 Oct 20 11:31] Self-test 28K passed!
[Worker #1 Oct 20 11:31] Test 1, 3200000 Lucas-Lehmer iterations of M172031 using FMA3 FFT length 8K, Pass1=128, Pass2=64, clm=2.
[Worker #16 Oct 20 11:31] Self-test 28K passed!
[Worker #16 Oct 20 11:31] Test 1, 3200000 Lucas-Lehmer iterations of M172031 using FMA3 FFT length 8K, Pass1=128, Pass2=64, clm=2.
[Worker #13 Oct 20 11:31] Self-test 28K passed!
[Worker #13 Oct 20 11:31] Test 1, 3200000 Lucas-Lehmer iterations of M172031 using FMA3 FFT length 8K, Pass1=128, Pass2=64, clm=2.
[Worker #9 Oct 20 11:31] Self-test 28K passed!
[Worker #9 Oct 20 11:31] Test 1, 3200000 Lucas-Lehmer iterations of M172031 using FMA3 FFT length 8K, Pass1=128, Pass2=64, clm=2.
[Worker #5 Oct 20 11:31] Self-test 28K passed!
[Worker #5 Oct 20 11:31] Test 1, 3200000 Lucas-Lehmer iterations of M172031 using FMA3 FFT length 8K, Pass1=128, Pass2=64, clm=2.
[Worker #3 Oct 20 11:31] Self-test 28K passed!
[Worker #3 Oct 20 11:31] Test 1, 3200000 Lucas-Lehmer iterations of M172031 using FMA3 FFT length 8K, Pass1=128, Pass2=64, clm=2.
[Worker #14 Oct 20 11:31] Self-test 28K passed!
[Worker #14 Oct 20 11:31] Test 1, 3200000 Lucas-Lehmer iterations of M172031 using FMA3 FFT length 8K, Pass1=128, Pass2=64, clm=2.
[Worker #10 Oct 20 11:31] Self-test 28K passed!
[Worker #10 Oct 20 11:31] Test 1, 3200000 Lucas-Lehmer iterations of M172031 using FMA3 FFT length 8K, Pass1=128, Pass2=64, clm=2.
[Worker #6 Oct 20 11:31] Self-test 28K passed!
[Worker #6 Oct 20 11:31] Test 1, 3200000 Lucas-Lehmer iterations of M172031 using FMA3 FFT length 8K, Pass1=128, Pass2=64, clm=2.
[Worker #4 Oct 20 11:31] Self-test 28K passed!
[Worker #4 Oct 20 11:31] Test 1, 3200000 Lucas-Lehmer iterations of M172031 using FMA3 FFT length 8K, Pass1=128, Pass2=64, clm=2.
[Worker #8 Oct 20 11:31] Self-test 28K passed!
[Worker #8 Oct 20 11:31] Test 1, 3200000 Lucas-Lehmer iterations of M172031 using FMA3 FFT length 8K, Pass1=128, Pass2=64, clm=2.
[Worker #7 Oct 20 11:31] Self-test 28K passed!
[Worker #7 Oct 20 11:31] Test 1, 3200000 Lucas-Lehmer iterations of M172031 using FMA3 FFT length 8K, Pass1=128, Pass2=64, clm=2.

[Worker #12 Oct 20 11:35] Self-test 8K passed!
[Worker #12 Oct 20 11:35] Test 1, 640000 Lucas-Lehmer iterations of M662593 using FMA3 FFT length 32K, Pass1=512, Pass2=64, clm=2.
[Worker #9 Oct 20 11:35] Self-test 8K passed!
[Worker #9 Oct 20 11:35] Test 1, 640000 Lucas-Lehmer iterations of M662593 using FMA3 FFT length 32K, Pass1=512, Pass2=64, clm=2.
[Worker #11 Oct 20 11:35] Self-test 8K passed!
[Worker #11 Oct 20 11:35] Test 1, 640000 Lucas-Lehmer iterations of M662593 using FMA3 FFT length 32K, Pass1=512, Pass2=64, clm=2.
[Worker #13 Oct 20 11:35] Self-test 8K passed!
[Worker #13 Oct 20 11:35] Test 1, 640000 Lucas-Lehmer iterations of M662593 using FMA3 FFT length 32K, Pass1=512, Pass2=64, clm=2.
[Worker #14 Oct 20 11:35] Self-test 8K passed!
[Worker #14 Oct 20 11:35] Test 1, 640000 Lucas-Lehmer iterations of M662593 using FMA3 FFT length 32K, Pass1=512, Pass2=64, clm=2.
[Worker #2 Oct 20 11:35] Self-test 8K passed!
[Worker #2 Oct 20 11:35] Test 1, 640000 Lucas-Lehmer iterations of M662593 using FMA3 FFT length 32K, Pass1=512, Pass2=64, clm=2.
[Worker #3 Oct 20 11:35] Self-test 8K passed!
[Worker #3 Oct 20 11:35] Test 1, 640000 Lucas-Lehmer iterations of M662593 using FMA3 FFT length 32K, Pass1=512, Pass2=64, clm=2.
[Worker #5 Oct 20 11:35] Self-test 8K passed!
[Worker #5 Oct 20 11:35] Test 1, 640000 Lucas-Lehmer iterations of M662593 using FMA3 FFT length 32K, Pass1=512, Pass2=64, clm=2.
[Worker #10 Oct 20 11:35] Self-test 8K passed!
[Worker #10 Oct 20 11:35] Test 1, 640000 Lucas-Lehmer iterations of M662593 using FMA3 FFT length 32K, Pass1=512, Pass2=64, clm=2.
[Worker #6 Oct 20 11:35] Self-test 8K passed!
[Worker #6 Oct 20 11:35] Test 1, 640000 Lucas-Lehmer iterations of M662593 using FMA3 FFT length 32K, Pass1=512, Pass2=64, clm=2.
[Worker #1 Oct 20 11:35] Self-test 8K passed!
[Worker #1 Oct 20 11:35] Test 1, 640000 Lucas-Lehmer iterations of M662593 using FMA3 FFT length 32K, Pass1=512, Pass2=64, clm=2.
[Worker #16 Oct 20 11:35] Self-test 8K passed!
[Worker #16 Oct 20 11:35] Test 1, 640000 Lucas-Lehmer iterations of M662593 using FMA3 FFT length 32K, Pass1=512, Pass2=64, clm=2.
[Worker #4 Oct 20 11:35] Self-test 8K passed!
[Worker #4 Oct 20 11:35] Test 1, 640000 Lucas-Lehmer iterations of M662593 using FMA3 FFT length 32K, Pass1=512, Pass2=64, clm=2.
[Worker #15 Oct 20 11:35] Self-test 8K passed!
[Worker #15 Oct 20 11:35] Test 1, 640000 Lucas-Lehmer iterations of M662593 using FMA3 FFT length 32K, Pass1=512, Pass2=64, clm=2.
[Worker #8 Oct 20 11:35] Self-test 8K passed!
[Worker #8 Oct 20 11:35] Test 1, 640000 Lucas-Lehmer iterations of M662593 using FMA3 FFT length 32K, Pass1=512, Pass2=64, clm=2.
[Worker #7 Oct 20 11:35] Self-test 8K passed!
[Worker #7 Oct 20 11:35] Test 1, 640000 Lucas-Lehmer iterations of M662593 using FMA3 FFT length 32K, Pass1=512, Pass2=64, clm=2.
[Worker #9 Oct 20 11:39] Self-test 32K passed!
[Worker #9 Oct 20 11:39] Test 1, 1800000 Lucas-Lehmer iterations of M250519 using FMA3 FFT length 12K, Pass1=256, Pass2=48, clm=1.
[Worker #13 Oct 20 11:39] Self-test 32K passed!
[Worker #13 Oct 20 11:39] Test 1, 1800000 Lucas-Lehmer iterations of M250519 using FMA3 FFT length 12K, Pass1=256, Pass2=48, clm=1.
[Worker #11 Oct 20 11:39] Self-test 32K passed!
[Worker #11 Oct 20 11:39] Test 1, 1800000 Lucas-Lehmer iterations of M250519 using FMA3 FFT length 12K, Pass1=256, Pass2=48, clm=1.
[Worker #6 Oct 20 11:39] Self-test 32K passed!
[Worker #6 Oct 20 11:39] Test 1, 1800000 Lucas-Lehmer iterations of M250519 using FMA3 FFT length 12K, Pass1=256, Pass2=48, clm=1.
[Worker #3 Oct 20 11:39] Self-test 32K passed!
[Worker #3 Oct 20 11:39] Test 1, 1800000 Lucas-Lehmer iterations of M250519 using FMA3 FFT length 12K, Pass1=256, Pass2=48, clm=1.
[Worker #12 Oct 20 11:39] Self-test 32K passed!
[Worker #12 Oct 20 11:39] Test 1, 1800000 Lucas-Lehmer iterations of M250519 using FMA3 FFT length 12K, Pass1=256, Pass2=48, clm=1.
[Worker #2 Oct 20 11:39] Self-test 32K passed!
[Worker #2 Oct 20 11:39] Test 1, 1800000 Lucas-Lehmer iterations of M250519 using FMA3 FFT length 12K, Pass1=256, Pass2=48, clm=1.
[Worker #5 Oct 20 11:39] Self-test 32K passed!
[Worker #5 Oct 20 11:39] Test 1, 1800000 Lucas-Lehmer iterations of M250519 using FMA3 FFT length 12K, Pass1=256, Pass2=48, clm=1.
[Worker #10 Oct 20 11:39] Self-test 32K passed!
[Worker #10 Oct 20 11:39] Test 1, 1800000 Lucas-Lehmer iterations of M250519 using FMA3 FFT length 12K, Pass1=256, Pass2=48, clm=1.
[Worker #14 Oct 20 11:39] Self-test 32K passed!
[Worker #14 Oct 20 11:39] Test 1, 1800000 Lucas-Lehmer iterations of M250519 using FMA3 FFT length 12K, Pass1=256, Pass2=48, clm=1.
[Worker #1 Oct 20 11:39] Self-test 32K passed!
[Worker #1 Oct 20 11:39] Test 1, 1800000 Lucas-Lehmer iterations of M250519 using FMA3 FFT length 12K, Pass1=256, Pass2=48, clm=1.
[Worker #16 Oct 20 11:39] Self-test 32K passed!
[Worker #16 Oct 20 11:39] Test 1, 1800000 Lucas-Lehmer iterations of M250519 using FMA3 FFT length 12K, Pass1=256, Pass2=48, clm=1.
[Worker #15 Oct 20 11:39] Self-test 32K passed!
[Worker #15 Oct 20 11:39] Test 1, 1800000 Lucas-Lehmer iterations of M250519 using FMA3 FFT length 12K, Pass1=256, Pass2=48, clm=1.
[Worker #7 Oct 20 11:39] Self-test 32K passed!
[Worker #7 Oct 20 11:39] Test 1, 1800000 Lucas-Lehmer iterations of M250519 using FMA3 FFT length 12K, Pass1=256, Pass2=48, clm=1.
[Worker #4 Oct 20 11:39] Self-test 32K passed!
[Worker #4 Oct 20 11:39] Test 1, 1800000 Lucas-Lehmer iterations of M250519 using FMA3 FFT length 12K, Pass1=256, Pass2=48, clm=1.
[Worker #8 Oct 20 11:39] Self-test 32K passed!
[Worker #8 Oct 20 11:39] Test 1, 1800000 Lucas-Lehmer iterations of M250519 using FMA3 FFT length 12K, Pass1=256, Pass2=48, clm=1.
C
Ĉ
^C
^Q
^C[Main thread Oct 20 11:40] Stopping all worker threads.
[Worker #8 Oct 20 11:40] Torture Test completed 3 tests in 12 minutes - 0 errors, 0 warnings.
[Worker #13 Oct 20 11:40] Torture Test completed 3 tests in 12 minutes - 0 errors, 0 warnings.
[Worker #13 Oct 20 11:40] Worker stopped.
[Worker #10 Oct 20 11:40] Torture Test completed 3 tests in 12 minutes - 0 errors, 0 warnings.
[Worker #10 Oct 20 11:40] Worker stopped.
[Worker #2 Oct 20 11:40] Torture Test completed 3 tests in 12 minutes - 0 errors, 0 warnings.
[Worker #2 Oct 20 11:40] Worker stopped.
[Worker #12 Oct 20 11:40] Torture Test completed 3 tests in 12 minutes - 0 errors, 0 warnings.
[Worker #12 Oct 20 11:40] Worker stopped.
[Worker #9 Oct 20 11:40] Torture Test completed 3 tests in 12 minutes - 0 errors, 0 warnings.
[Worker #9 Oct 20 11:40] Worker stopped.
[Worker #5 Oct 20 11:40] Torture Test completed 3 tests in 12 minutes - 0 errors, 0 warnings.
[Worker #5 Oct 20 11:40] Worker stopped.
[Worker #6 Oct 20 11:40] Torture Test completed 3 tests in 12 minutes - 0 errors, 0 warnings.
[Worker #6 Oct 20 11:40] Worker stopped.
[Worker #3 Oct 20 11:40] Torture Test completed 3 tests in 12 minutes - 0 errors, 0 warnings.
[Worker #3 Oct 20 11:40] Worker stopped.
[Worker #14 Oct 20 11:40] Torture Test completed 3 tests in 12 minutes - 0 errors, 0 warnings.
[Worker #14 Oct 20 11:40] Worker stopped.
[Worker #11 Oct 20 11:40] Torture Test completed 3 tests in 12 minutes - 0 errors, 0 warnings.
[Worker #11 Oct 20 11:40] Worker stopped.
[Worker #7 Oct 20 11:40] Torture Test completed 3 tests in 12 minutes - 0 errors, 0 warnings.
[Worker #7 Oct 20 11:40] Worker stopped.
[Worker #4 Oct 20 11:40] Torture Test completed 3 tests in 12 minutes - 0 errors, 0 warnings.
[Worker #4 Oct 20 11:40] Worker stopped.
[Worker #16 Oct 20 11:40] Torture Test completed 3 tests in 12 minutes - 0 errors, 0 warnings.
[Worker #16 Oct 20 11:40] Worker stopped.
[Worker #1 Oct 20 11:40] Torture Test completed 3 tests in 12 minutes - 0 errors, 0 warnings.
[Worker #1 Oct 20 11:40] Worker stopped.
[Worker #8 Oct 20 11:40] Worker stopped.
[Worker #15 Oct 20 11:40] Torture Test completed 3 tests in 12 minutes - 0 errors, 0 warnings.
[Worker #15 Oct 20 11:40] Worker stopped.
[Main thread Oct 20 11:40] Execution halted.
             Main Menu

         1.  Test/Primenet
         2.  Test/Worker threads
         3.  Test/Status
         4.  Test/Continue
         5.  Test/Exit
         6.  Advanced/Test
         7.  Advanced/Time
         8.  Advanced/P-1
         9.  Advanced/ECM
        10.  Advanced/Manual Communication
        11.  Advanced/Unreserve Exponent
        12.  Advanced/Quit Gimps
        13.  Options/CPU
        14.  Options/Preferences
        15.  Options/Torture Test
        16.  Options/Benchmark
        17.  Help/About
        18.  Help/About PrimeNet Server
Your choice: 
```

```
 for i in {1..10}; do sleep 30; echo "After $((i*30)) seconds"; sudo liquidctl status; done
After 30 seconds
[sudo] Mot de passe de olivier : 
Device 0, Corsair RM650i (experimental)
WARNING: reading the +12V OCP mode is an experimental feature
Current uptime                    0:47:06  
Total uptime            13 days, 13:47:06  
Temperature 1                        41.2  °C
Temperature 2                        32.8  °C
Fan control mode                 Software  
…